remove comment policy

+++ NexTube.Application/CQRS/Comments/VideoComments/Commands/DeleteComment/DeleteCommentCommandHandler.cs
@@ -1,11 +1,13 @@
 ﻿using Ardalis.GuardClauses;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 using NexTube.Application.Common.DbContexts;
 using NexTube.Application.Common.Interfaces;
 using NexTube.Application.Common.Models;
 using NexTube.Application.CQRS.Comments.VideoComments.Commands.AddComment;
 using NexTube.Domain.Entities;
 using System.ComponentModel.Design;
+using WebShop.Application.Common.Exceptions;
 
 namespace NexTube.Application.CQRS.Comments.VideoComments.Commands.DeleteComment {
     public class DeleteCommentCommandHandler : IRequestHandler<DeleteCommentCommand, Unit> {
@@ -16,10 +18,16 @@ namespace NexTube.Application.CQRS.Comments.VideoComments.Commands.DeleteComment
         }
 
         public async Task<Unit> Handle(DeleteCommentCommand request, CancellationToken cancellationToken) {
-            var comment = await _dbContext.VideoComments.FindAsync(request.CommentId);
+            var comment = await _dbContext.VideoComments
+                .Include(vc => vc.Creator)
+                .Where(vc => vc.Id == request.CommentId)
+                .FirstOrDefaultAsync();
 
             if (comment is null)
                 throw new NotFoundException(request.CommentId.ToString(), nameof(VideoCommentEntity));
+
+            if (comment?.Creator?.Id != request?.Requester?.Id)
+                throw new ForbiddenAccessException();
 
             _dbContext.VideoComments.Remove(comment);
